Переписан процессор диплома
